// sdk/multisig/src/generated/types/MutateCompressedMultisigArgs.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
import { CompressedProof, compressedProofBeet } from './CompressedProof'
import { LightMultisigData, lightMultisigDataBeet } from './LightMultisigData'
import {
  PackedMerkleContext,
  packedMerkleContextBeet,
} from './PackedMerkleContext'
export type MutateCompressedMultisigArgs = {
  compressedProof: CompressedProof
  address: number[] /* size: 32 */
  multisigData: LightMultisigData
  merkleTreeRootIndex: number
  merkleContext: PackedMerkleContext
}

/**
 * @category userTypes
 * @category generated
 */
export const mutateCompressedMultisigArgsBeet =
  new beet.FixableBeetArgsStruct<MutateCompressedMultisigArgs>(
    [
      ['compressedProof', compressedProofBeet],
      ['address', beet.uniformFixedSizeArray(beet.u8, 32)],
      ['multisigData', lightMultisigDataBeet],
      ['merkleTreeRootIndex', beet.u16],
      ['merkleContext', packedMerkleContextBeet],
    ],
    'MutateCompressedMultisigArgs'
  )