// sdk/multisig/src/generated/types/PackedMerkleContext.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
import { QueueIndex, queueIndexBeet } from './QueueIndex'
export type PackedMerkleContext = {
  merkleTreePubkeyIndex: number
  nullifierQueuePubkeyIndex: number
  leafIndex: number
  queueIndex: beet.COption<QueueIndex>
}

/**
 * @category userTypes
 * @category generated
 */
export const packedMerkleContextBeet =
  new beet.FixableBeetArgsStruct<PackedMerkleContext>(
    [
      ['merkleTreePubkeyIndex', beet.u8],
      ['nullifierQueuePubkeyIndex', beet.u8],
      ['leafIndex', beet.u32],
      ['queueIndex', beet.coption(queueIndexBeet)],
    ],
    'PackedMerkleContext'
  )